+++ src/Command/ResetAllSettingsCommand.php
@@ -155,7 +155,7 @@ class ResetAllSettingsCommand extends Command
 
             ['name' => SettingName::CAPPORT_ENABLED->value, 'value' => 'false'],
             ['name' => SettingName::CAPPORT_PORTAL_URL->value, 'value' => 'https://example.com/'],
-            ['name' => SettingName::CAPPORT_VENUE_INFO_URL->value, 'value' => ' https://openroaming.org/'],
+            ['name' => SettingName::CAPPORT_VENUE_INFO_URL->value, 'value' => 'https://openroaming.org/'],
 
             ['name' => SettingName::SMS_USERNAME->value, 'value' => ''],
             ['name' => SettingName::SMS_USER_ID->value, 'value' => ''],
@@ -178,7 +178,11 @@ class ResetAllSettingsCommand extends Command
             ['name' => SettingName::FREERADIUS_LAST_CONNECTION_CRON->value, 'value' => '* 3 * * *'],
             ['name' => SettingName::DOMAIN_BLACKLIST_IMPORT_CRON->value, 'value' => '0 4 * * *'],
             ['name' => SettingName::CRON_ADVANCED_STATUS->value, 'value' => 'OFF'],
+            ['name' => SettingName::CLOUDFLARE_TOKEN->value, 'value' => ''],
             ['name' => SettingName::ENABLE_RADIUS_TLS_RESET->value, 'value' => 'true'],
+            ['name' => SettingName::RETURN_APPS_ENABLED->value, 'value' => 'OFF'],
+            ['name' => SettingName::RETURN_APPS_PACKAGE_NAME_ANDROID->value, 'value' => 'EditMe'],
+            ['name' => SettingName::RETURN_APPS_ID_IOS->value, 'value' => 'EditMe.EditMe'],
         ];
 
         // phpcs:disable Generic.Files.LineLength.TooLong