+++ src/DTO/UserUpdateDTO.php
@@ -59,6 +59,7 @@ class UserUpdateDTO
     public PermissionLevel $termsPolicies = PermissionLevel::NONE;
     public PermissionLevel $cronSchedule = PermissionLevel::NONE;
     public PermissionLevel $certificatesManagement = PermissionLevel::NONE;
+    public PermissionLevel $returnAppsManagement = PermissionLevel::NONE;
     public PermissionLevel $authenticationMethods = PermissionLevel::NONE;
     public PermissionLevel $twoFactorAuth = PermissionLevel::NONE;
     public PermissionLevel $ldapSynchronization = PermissionLevel::NONE;
@@ -164,6 +165,7 @@ class UserUpdateDTO
         'termsPolicies' => 'TERMS_POLICIES',
         'cronSchedule' => 'CRON_SCHEDULE',
         'certificatesManagement' => 'CERTIFICATES_MANAGEMENT',
+        'returnAppsManagement' => 'RETURN_APPS_MANAGEMENT',
         'authenticationMethods' => 'AUTHENTICATION_METHODS',
         'twoFactorAuth' => 'TWO_FACTOR_AUTH',
         'ldapSynchronization' => 'LDAP_SYNCHRONIZATION',